Move ddl to base_sqlalchemy
