+++ ibis/backends/impala/ddl.py
@@ -15,8 +15,8 @@
 import json
 
 from ibis.backends.base.sql import type_to_sql_string
-from ibis.backends.base_sql import ddl as base_sql_ddl
-from ibis.backends.base_sql.ddl import (
+from ibis.backends.base_sqlalchemy import ddl as base_sql_ddl
+from ibis.backends.base_sqlalchemy.ddl import (
     AlterTable,
     BaseDDL,
     CreateTable,
