+++ ibis/backends/impala/kudu_support.py
@@ -4,7 +4,7 @@ import kudu
 import pandas as pd
 
 import ibis.expr.datatypes as dt
-from ibis.backends.base_sql.ddl import (
+from ibis.backends.base_sqlalchemy.ddl import (
     CreateTable,
     format_schema,
     format_tblproperties,
